refactor(web_core): streamline topology validation and componentRefMap
- Remove redundant analyzeTopology and A2uiValidator helpers in favor of SurfaceComponentsModel topology validation.
- Add lazy componentRefMap property to Catalog class and consume directly in SurfaceComponentsModel.getChildReferences(), eliminating external WeakMap caches.
- Extract reference map utilities and unit tests to catalog/reference-map.ts and catalog/reference-map.test.ts.
- Consolidate validation configuration and integrity checks in validating/integrity-checker.ts and validating/integrity-checker.test.ts.
- Update Angular and React test suites to provide the required catalog parameter when constructing ComponentModel instances.
