+++ typescript/web_core/src/v1_0/index.ts
@@ -18,4 +18,4 @@ export * from './schema/index.js';
 export * from './functions/system_functions.js';
 export * from './functions/validation_functions_api.js';
 export * from './functions/validation_functions.js';
-export * from '../validating/index.js';
+export * from '../validating/integrity-checker.js';
