--- typescript/web_core/src/validating/validator.ts
@@ -1,129 +1,0 @@
-/*
- * Copyright 2024 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import {A2uiValidationError} from '../errors.js';
-import {ProtocolVersion} from '../processing/adapters/base.js';
-import {VersionAdapterFactory} from '../processing/adapters/factory.js';
-import {formatZodIssue} from '../processing/message-processor.js';
-import {
-  CatalogOrRefMapInput,
-  IntegrityOptions,
-  validateComponentIntegrity,
-  validateRecursionAndPaths,
-} from './integrity-checker.js';
-import {analyzeTopology, TopologyOptions} from './topology-analyzer.js';
-
-/** Combined configuration specifying integrity, topology, version, and catalog validation rules. */
-export interface ValidationConfig extends IntegrityOptions, TopologyOptions {
-  /** Target protocol version expected for incoming messages (e.g. 'v0.8', 'v0.9', 'v1.0'). */
-  targetVersion?: ProtocolVersion | string;
-  /** When false, verifies that all component types exist in the surface catalog. Default: false. */
-  allowUnknownElements?: boolean;
-  /** Allowed top-level message operation types (e.g. ['createSurface', 'updateComponents']). */
-  allowedMessages?: string[];
-}
-
-/** Strict validation configuration requiring root node presence, no orphans, valid references, and catalog compliance. */
-export const STRICT_VALIDATION: ValidationConfig = Object.freeze({
-  allowOrphanComponents: false,
-  allowDanglingReferences: false,
-  allowMissingRoot: false,
-  allowUnknownElements: false,
-});
-
-/** Relaxed validation configuration permitting orphan components, missing root, dangling references, and unknown elements. */
-export const RELAXED_VALIDATION: ValidationConfig = Object.freeze({
-  allowOrphanComponents: true,
-  allowDanglingReferences: true,
-  allowMissingRoot: true,
-  allowUnknownElements: true,
-});
-
-/** Options for fine-tuning component validation operations. */
-export interface ValidateComponentsOptions {
-  /** Whether to bypass path syntax and depth recursion checks on component objects. */
-  skipRecursionCheck?: boolean;
-}
-
-/**
- * High-level validator for auditing A2UI message streams, components, and graph topology.
- *
- * @example
- * ```ts
- * const validator = new A2uiValidator();
- * validator.validate(messages, catalog, STRICT_VALIDATION);
- * ```
- */
-export class A2uiValidator {
-  /**
-   * Validates a list of protocol messages against Zod message envelope schemas.
-   *
-   * @param messages Stream of raw message objects to audit.
-   * @throws {A2uiValidationError} If any message fails Zod envelope schema validation.
-   */
-  public validateProtocolEnvelope(messages: Array<Record<string, any>>): void {
-    if (!Array.isArray(messages)) {
-      throw new A2uiValidationError('Message stream must be an array of objects');
-    }
-
-    for (let idx = 0; idx < messages.length; idx++) {
-      const msg = messages[idx];
-      if (typeof msg !== 'object' || msg === null) {
-        throw new A2uiValidationError(`Message must be an object at index ${idx}`);
-      }
-
-      let adapter;
-      try {
-        adapter = VersionAdapterFactory.resolveFromPayload(msg);
-      } catch {
-        adapter = VersionAdapterFactory.getAdapter('v1.0');
-      }
-
-      const parseResult = (adapter as any).schema.safeParse(msg);
-      if (!parseResult.success) {
-        const formatted = parseResult.error.errors.map(formatZodIssue).join('; ');
-        throw new A2uiValidationError(
-          `Validation failed for message at index ${idx}: ${formatted}`,
-          parseResult.error.issues,
-        );
-      }
-    }
-  }
-
-  /**
-   * Validates component list integrity and graph topology.
-   *
-   * @param components List of component definition objects.
-   * @param catalogOrRefMap Reference field definitions per component type, Catalog instance, or list/map of Catalogs.
-   * @param config Validation settings for integrity and topology.
-   * @param options Additional component validation options.
-   * @throws {A2uiIntegrityError} If integrity check or reachability check fails.
-   * @throws {A2uiRecursionError} If graph recursion or self-reference is found.
-   * @throws {A2uiValidationError} If invalid path syntax is encountered.
-   */
-  public validateComponents(
-    components: Array<Record<string, any>>,
-    catalogOrRefMap: CatalogOrRefMapInput,
-    config: ValidationConfig = STRICT_VALIDATION,
-    options: ValidateComponentsOptions = {},
-  ): void {
-    if (!options.skipRecursionCheck) {
-      validateRecursionAndPaths(components);
-    }
-    validateComponentIntegrity(components, catalogOrRefMap, config);
-    analyzeTopology(components, catalogOrRefMap, config);
-  }
-}
